+++ Source/IO/WinCopies.GUI.IO.Shared/ObjectModel/BrowsableObjectInfoViewModel.Partial/Browsability.Partial/Browsability.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.ObjectModel;
-using WinCopies.IO.ObjectModel;
-
-namespace WinCopies.GUI.IO.ObjectModel
+﻿namespace WinCopies.GUI.IO.ObjectModel
 {
     public partial class BrowsableObjectInfoViewModel
     {
-        private partial struct Browsability
+        private partial struct _Browsability
         {
-
             public Items _items;
             public Parent _parent;
         }