+++ Source/IO/WinCopies.GUI.Shell.Shared/Controls/FileSystemDialogBoxContent.cs
@@ -4,7 +4,7 @@ using System.Windows.Controls;
 
 using WinCopies.GUI.IO.ObjectModel;
 using WinCopies.IO.ObjectModel;
-
+using WinCopies.Util.Data;
 using static WinCopies.Util.Desktop.UtilHelpers;
 
 namespace WinCopies.GUI.Shell
@@ -24,14 +24,18 @@ namespace WinCopies.GUI.Shell
 
         private static DependencyProperty Register<T>(in string propertyName, in PropertyMetadata metadata) => Register<T, FileSystemDialogBoxContent>(propertyName, metadata);
 
-        public static readonly DependencyProperty FilterProperty = Register<System.Collections.Generic.IEnumerable<string>>(nameof(Filter), new PropertyMetadata((DependencyObject d, DependencyPropertyChangedEventArgs e) =>
+        public static readonly DependencyProperty FiltersProperty = Register<System.Collections.Generic.IEnumerable<INamedObject<string>>>(nameof(Filters), new PropertyMetadata((DependencyObject d, DependencyPropertyChangedEventArgs e) =>
         {
             if (((FileSystemDialogBoxContent)d).Mode == FileSystemDialogBoxMode.SelectFolder)
 
-                throw new InvalidOperationException($"{nameof(Mode)} must not be set to {nameof(FileSystemDialogBoxMode.SelectFolder)} to modify {nameof(Filter)}.");
+                throw new InvalidOperationException($"{nameof(Mode)} must not be set to {nameof(FileSystemDialogBoxMode.SelectFolder)} to modify {nameof(Filters)}.");
         }));
 
-        public System.Collections.Generic.IEnumerable<string> Filter { get => (System.Collections.Generic.IEnumerable<string>)GetValue(FilterProperty); set => SetValue(FilterProperty, value); }
+        public System.Collections.Generic.IEnumerable<INamedObject<string>> Filters { get => (System.Collections.Generic.IEnumerable<INamedObject<string>>)GetValue(FiltersProperty); set => SetValue(FiltersProperty, value); }
+
+        public static readonly DependencyProperty SelectedFilterProperty = Register<INamedObject<string>>(nameof(SelectedFilter));
+
+        public INamedObject<string> SelectedFilter { get => (INamedObject<string>)GetValue(FiltersProperty); set => SetValue(FiltersProperty, value); }
 
         public static readonly DependencyProperty PredicateProperty = Register<Predicate<IBrowsableObjectInfo>>(nameof(Predicate));
 
@@ -47,9 +51,9 @@ namespace WinCopies.GUI.Shell
 
         public static readonly DependencyProperty ModeProperty = Register<FileSystemDialogBoxMode>(nameof(Mode), new PropertyMetadata((DependencyObject d, DependencyPropertyChangedEventArgs e) =>
         {
-            if ((FileSystemDialogBoxMode)e.NewValue == FileSystemDialogBoxMode.SelectFolder && ((FileSystemDialogBoxContent)d).Filter != null)
+            if ((FileSystemDialogBoxMode)e.NewValue == FileSystemDialogBoxMode.SelectFolder && ((FileSystemDialogBoxContent)d).Filters != null)
 
-                throw new InvalidOperationException($"Cannot set {nameof(Mode)} to {nameof(FileSystemDialogBoxMode.SelectFolder)} when {nameof(Filter)} has value.");
+                throw new InvalidOperationException($"Cannot set {nameof(Mode)} to {nameof(FileSystemDialogBoxMode.SelectFolder)} when {nameof(Filters)} has value.");
         }));
 
         public FileSystemDialogBoxMode Mode { get => (FileSystemDialogBoxMode)GetValue(ModeProperty); set => SetValue(ModeProperty, value); }