+++ Source/IO/WinCopies.GUI.Shell.Shared/ViewModels/FileSystemDialogBoxViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 using WinCopies.GUI.IO.ObjectModel;
 using WinCopies.IO.ObjectModel;
@@ -9,7 +10,9 @@ namespace WinCopies.GUI.Shell
 {
     public class FileSystemDialogBoxViewModel : ViewModel<IFileSystemDialog>, IFileSystemDialog
     {
-        public System.Collections.Generic.IEnumerable<string> Filter { get => ModelGeneric.Filter; set { ModelGeneric.Filter = value; OnPropertyChanged(nameof(Filter)); } }
+        public System.Collections.Generic.IEnumerable<INamedObject<string>> Filters { get => ModelGeneric.Filters; set { ModelGeneric.Filters = value; OnPropertyChanged(nameof(Filters)); } }
+
+        public INamedObject<string> SelectedFilter { get => ModelGeneric.SelectedFilter; set { ModelGeneric.SelectedFilter = value; OnPropertyChanged(nameof(Filters)); } }
 
         public Predicate<IBrowsableObjectInfo> Predicate { get => ModelGeneric.Predicate; set { ModelGeneric.Predicate = value; OnPropertyChanged(nameof(Predicate)); } }
 
@@ -22,5 +25,23 @@ namespace WinCopies.GUI.Shell
         public FileSystemDialogBoxViewModel(in IFileSystemDialog model) : base(model) { /* Left empty. */ }
 
         public FileSystemDialogBoxViewModel(in FileSystemDialogBoxMode mode) : this(new FileSystemDialog(mode)) { /* Left empty. */ }
+
+        public bool CompletePredicate(IBrowsableObjectInfo browsableObjectInfo)
+        {
+            if (SelectedFilter != null)
+            {
+                string filter = SelectedFilter.Value;
+
+                if (!WinCopies.IO.Path.Match(browsableObjectInfo.Name, filter))
+
+                    return false;
+            }
+
+            if (Predicate == null)
+
+                return true;
+
+            return Predicate(browsableObjectInfo);
+        }
     }
 }