+++ Source/IO/WinCopies.IO.Shell.Shared/Guids.cs
@@ -1,0 +1,17 @@
+﻿namespace WinCopies.IO.Guids.Shell
+{
+    public static class Process
+    {
+        public static class Shell
+        {
+            public const string Copy = "084ff8d5-c66e-40bc-8ea4-7fa2dd30bd21";
+            public const string Deletion = "af5a27a6-dda3-49f3-b287-cfe6f070a812";
+        }
+
+        public static class Archive
+        {
+            public const string Compression = "02d03025-8cec-49b7-a639-46b9dcb2569b";
+            public const string Extraction = "96272e6c-be30-4db8-b48f-54b6fb425302";
+        }
+    }
+}